requirements.txt
sanic

pytest==7.4.0
PyJWT==2.8.0

sanic-testing
pytest-asyncio==0.21.1

pytest-cover~=3.0.0
flake8==6.1.0
mypy==1.5.1
black==23.7.0
pylint==2.17.5
isort==5.12.0

types-python-dateutil
